requirements.txt
streamlit>=1.36
openai>=1.40
psycopg[binary]>=3.1
pandas>=2.2
python-dotenv>=1.0
langchain-openai>=0.2
langchain-core>=0.2
